Backend/requirements.txt
browser-use
playwright
openai>=1.0.0
langchain-openai
cartesia
elevenlabs>=1.0.0
websockets>=12.0
requests>=2.31
numpy>=1.24
sounddevice>=0.4.6
soundfile
python-dotenv>=1.0
fastapi
uvicorn






# ── Core bot dependencies ─────────────────────────────────────────────────────
fastapi
uvicorn[standard]
python-dotenv
playwright
openai
requests
numpy
pydub
pyaudio

# ── Cartesia TTS ──────────────────────────────────────────────────────────────
cartesia

# ── Google Calendar API (for auto-creating Meet links) ────────────────────────
google-api-python-client
google-auth-httplib2
google-auth-oauthlib